requirements.txt
pandas
numpy
matplotlib
opencv-python
scikit-learn
librosa
ffmpeg-python
seaborn 
umap
scenedetect